requirements.txt
flask>=3.0.0
flask-wtf>=1.2.0
flask-limiter>=3.5.0
openai>=1.0.0
python-dotenv>=1.0.0
flask-dance[google]>=7.0.0
requests>=2.28.0
tenacity>=8.0.0
gunicorn>=22.0.0
Pillow>=10.0.0
pytest>=8.0.0
